requirements.txt
# FastAPI and Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0

# Google Cloud
google-cloud-storage==2.10.0
google-cloud-pubsub==2.18.4
google-cloud-sql-connector==1.4.3

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# AI/LLM
openai==1.3.0
langchain==0.0.335
langchain-google-vertexai==0.0.6

# Static Analysis Tools
pylint==3.0.2
bandit==1.7.5
radon==6.0.1
flake8==6.1.0
mypy==1.7.1

# Utilities
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dateutil==2.8.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Monitoring
prometheus-client==0.19.0
